Move 'FeedImagePresenter' and collaborators to production
